feat: Update useFDAOEmissionsAPR to multiply APR by 100
This commit modifies the useFDAOEmissionsAPR hook to multiply the calculated APR by 100. This change ensures that the APR value is represented as a percentage. The updated calculation will provide a more accurate representation of the emissions APR for the given pool.

Commit message follows the conventional format of "feat: <description>".